+++ src/Model/BannerCategoryModel.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (c) 2020 Heimrich & Hannot GmbH
+ *
+ * @author  Thomas Körner <t.koerner@heimrich-hannot.de>
+ * @license http://www.gnu.org/licences/lgpl-3.0.html LGPL
+ */
+
+
+namespace HeimrichHannot\BannerPlusBundle\Model;
+
+
+use Contao\Model;
+
+class BannerCategoryModel extends Model
+{
+    protected static $strTable = 'tl_banner_category';
+}