+++ src/Template/MultipleBannerTemplate.php
@@ -14,6 +14,7 @@ namespace HeimrichHannot\BannerPlusBundle\Template;
 
 use BugBuster\Banner\BannerMultiple;
 use HeimrichHannot\BannerPlusBundle\Model\BannerModel;
+use HeimrichHannot\MediaQuery\Viewport;
 
 class MultipleBannerTemplate extends BannerMultiple
 {
@@ -50,5 +51,14 @@ class MultipleBannerTemplate extends BannerMultiple
         $this->Template->banners = $arrBanners;
         return $this->Template;
     }
+
+    protected function setStatViewUpdate($arrBannerData, $module_id, $banner_useragent)
+    {
+        if (!SingleBannerTemplate::isVisibleBanner($this->arrCategoryValues['id'])) {
+            return;
+        }
+        parent::setStatViewUpdate($arrBannerData, $module_id, $banner_useragent);
+    }
+
 
 }